+++ packages/legend-query-builder/src/components/fetch-structure/QueryBuilderPostFilterPanel.tsx
@@ -95,7 +95,7 @@ import {
 import {
   type QueryBuilderWindowColumnDragSource,
   QUERY_BUILDER_WINDOW_COLUMN_DND_TYPE,
-} from '../../stores/fetch-structure/tds/olapGroupBy/QueryBuilderWindowState.js';
+} from '../../stores/fetch-structure/tds/window/QueryBuilderWindowState.js';
 import type { QueryBuilderTDSColumnState } from '../../stores/fetch-structure/tds/QueryBuilderTDSColumnState.js';
 
 const QueryBuilderPostFilterConditionContextMenu = observer(
