+++ packages/legend-query-builder/src/stores/fetch-structure/tds/window/QueryBuilderWindowGroupByOperatorLoader.ts
