+++ packages/legend-query-builder/src/stores/fetch-structure/tds/window/QueryBuilderWindowStateBuilder.ts
