+++ packages/legend-query-builder/src/stores/fetch-structure/tds/window/QueryBuilderWindowValueSpecificationBuilder.ts
