+++ packages/legend-query-builder/src/stores/fetch-structure/tds/window/operators/QueryBuilderTDS_WindowOperator_Min.ts
